feat: refactor to start adding more componentes
