+++ packages/asset-packs/src/versioning/components/index.ts
@@ -1,0 +1,1 @@
+export * from './counter';
